move common ArtifactTool files to packaging-common
