+++ Tasks/Common/packaging-common/universal/Authentication.ts
