+++ Tasks/UniversalPackagesV0/package.json
@@ -17,7 +17,7 @@
     },
     "homepage": "https://github.com/Microsoft/vsts-tasks#readme",
     "dependencies": {
-        "adm-zip": "^0.4.11",
+        "@types/mocha": "^5.2.5",
         "packaging-common": "file:../../_build/Tasks/Common/packaging-common-1.0.1.tgz",
         "utility-common": "file:../../_build/Tasks/Common/utility-common-1.0.2.tgz",
         "vso-node-api": "6.5.0",
