+++ Tasks/UniversalPackagesV0/task.json
@@ -10,7 +10,7 @@
     "version": {
         "Major": 0,
         "Minor": 147,
-        "Patch": 0
+        "Patch": 1
     },
     "runsOn": [
         "Agent",
