+++ Tasks/UniversalPackagesV0/task.loc.json
@@ -4,12 +4,13 @@
   "friendlyName": "ms-resource:loc.friendlyName",
   "description": "ms-resource:loc.description",
   "author": "Microsoft Corporation",
+  "helpUrl": "https://aka.ms/universalpackagesannounce",
   "helpMarkDown": "ms-resource:loc.helpMarkDown",
   "category": "Package",
   "version": {
     "Major": 0,
     "Minor": 147,
-    "Patch": 0
+    "Patch": 1
   },
   "runsOn": [
     "Agent",
